DeleteUnprocessedSong UseCase works with tests

+++ lib/pl2/lib/pl/use-cases/delete_unprocessed_song.rb
@@ -2,8 +2,12 @@ module PL
   class DeleteUnprocessedSong < UseCase
     def run(key)
       handler = PL::AudioFileStorageHandler.new
-      handler.delete_unprocessed_song(key)
-      return success
+      if handler.unprocessed_song_exists?(key)
+        handler.delete_unprocessed_song(key)
+        return success
+      else
+        return failure :unprocessed_song_not_found
+      end
     end
   end
 end
